Se integro con el endpoint de Unirse a Partida. Faltan Tests
