+++ src/main.jsx
@@ -1,14 +1,14 @@
-import "./Style/index.css"
+import "./index.css"
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import Listar_Partidas from './Listar_Partidas.jsx'
+import Listar_Partidas from './components/listar_partida/Listar_Partidas.jsx'
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path='/Lobby' element={<Listar_Partidas />}/>
+        <Route path='/Lobby' element={<Listar_Partidas jugador_id={"sdsda"} />} />
         <Route path="/Partida/:id" />
       </Routes>
     </BrowserRouter>
